feat: client daemon PoC
A PoC of a client daemon to sent "hearbeats" to Prometheus Server

Implemented parts:

* 2 modes of running:
** daemon - to do the heartbeat
** once - to do one heartbeat and end (e.g. for testing)

possible to configure (mostly for testing):
- tick - interval of hearbeats in seconds
- prometheus Url - where to do http request
- host cert path - which cert to use for loading Host ID and for MTLS
  auth
- host cert key path - for MTLS auth

the server also:
- reloads data on SIGHUP
- stops on SIGTERM, SIGINT

This is a PoC, so polishing expected. Also this is my first program in
golang os feedback welcome.

Actual contacting of Prometheus Server was not yet tested.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>