fix: Print debug messages by default without further configuration
The initial instance of the logger that is created by default should have its
log level set to DEBUG to be able to see debug messages from unit tests.

Milton changes this log level based on the configuration or fallbacks to INFO,
so this change has no impact on the usual run of the client tool.