feat(notify): recreate http client on host info change
Http client needs to be recreated when certificate changes. PrometheusNotifier created
another detection mechanism by checking if host id changed. This fails
e.g., in cases when certificate is renewed (id is the same).

This patch changes the behavior to more generic - to reload on any hostinfo change (host
info has to be reloaded when certifcate changes) thus transitively also
on the certificate change.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>