fix: tune http Transport config
Set's other defaults for transport which were otherwise lost by not
using the default transport.

`IdleConnTimeout`, `TLSHandshakeTimeout` and `ExpectContinueTimeout` are
the same as in `DefaultTransport`.

`MaxIdleConns` was set to 2 instead of default 100 just reduce it as we
expect one connection to a server at a time anyway and also
`DefaultMaxIdleConnsPerHost` is 2. I.e. it probably doesn't have
practical effect.

But it is probably useful to have keep-alive configuration for possible
re-tries when remote write returns 5xx.

DialContext and ForceAttemptHTTP2 were not set.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>