fix: throw away samples on Prometheus Remote Write 4xx errors (#15)
This change adds concept of Recoverable and Non-Recoverable error in
Notifier interface. Deamon throws away the samples on non-recoverable
error.

PrometheusNotifier was modified to produce Non-recoverable error on 4xx
http error family and on unknown http errors. This should be in complience
with prometheus remote write spec as the client should not retry on 4xx
except on 429.

This also means that host-metering will throw away samples from time
when
- it was misconfigured with wrong URL (404)
- auth issue happened (403)

It doesn't throw away samples on connectivity issues (no response
from server).

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>