fix: dropping of expired samples (#19)
* fix: log dropping of expired samples on INFO level

So that it is clearly visible even on INFO level - as this should be
unusual situation.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>

* fix: dropping of expired samples

Dropping of expired samples did not work correctly. It was truncating to
index: "lastIndex - number of non-expired samples" but this  might not
be the correct index as samples are often interlanced with checkpoints,
e.g. imagine wal:

 0: checkpoint
 1: sample 1 (now expired)
 2: checkpoint
 3: sample 2 (now expired)
 4: checkpoint
 5: sample 3
 6: checkpoint
 7: sample 4
 8: checkpoint
 9: sample 5
10: checkpoint
11: sample 6
12: checkpoint
13: sample 7
14: checkpoint:

removal of expired would be: 14 - 5 = 9, as:
- checkpoint: 14
- number of non-expired samples: 5

i.e. on recoverable failure it would drop samples up to sample 5(index
9) but it should drop only upto to index after sample 2 (index 4).

This patch in introducing a new method which traverses WAL and finds to
index after number of samples to remove - to behave as expected.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>

---------

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>