test: fix daemon test compilation issue
PR with commit 5f5e552abb5de46981fa9324341850b41eb1e937 was created
before commit 6cf0df3506aee8819aa9e1abd4c7bb4f38c3e23e but the latter
commit was merged first and the first was not rebased. The former changed
data type of `Product` which was used by latter. Thus tests introduced by
latter could not be compiled.

This fix is changing the tests from the latter to use a different field
so that it doesn't need to compare slices and thus fixing the datatype issue.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>