build: do not use %gopkg
The macro is not defined on RHEL/CentOS/EL 7.

The original idea was to wrap it with some `if 0%{?rhel} > 7 || 0%{?fedora}`
but this was failing in `packit srpm`.

The %gopkg is producing the host-metering-debugsource package, this is
not critical and thus is was removed. Individual downstreams can add it
if needed.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>